gpu: Move relevant things to gpuav_instrumentation
